refactor: extract shared opencode helpers and add PR identifier validation
Extract common types, utilities, schemas, and step functions from
implement-loop.ts and review-loop.ts into opencode-helpers.ts to
eliminate DRY violations. Add input format validation for PR identifier
in review-loop.ts.
